refactor(consensus): name execution driver -> application and use actor pattern (#575)
This patch renames the at times confusing execution driver to
application. It also reorganizes the crate to follow a similar pattern
as commonware and the recently introduced dkg and epoch managers:

+ the mailbox is in `application::ingress`
+ the actor is in `application::actor`
+ the config is found at `application::Config` (and
`ExecutionDriverBuilder` renamed to `Config`).
+ the former `ExecutionDriver` is now named `application::Actor`
+ the former `ExecutionDriverMailbox` is now named
`application::Mailbox`.

In a followup, the `executor` actor will also be restructured, and the
entire application placed at the top level of the library (instead of
under the `consensus` module).

A followup will also address the unnecessarily complex generics on the
application actor that are supposed to signal its initialization state.
These will be hidden as an implementation details.