feat(revm): Update fee token selection logic (#1022)
Closes #992
Closes  #993

Stacks on #1015 

This PR updates `user_or_tx_fee_token()` to handle if `tx.to` is the
`StablecoinExchange`. Additionally this PR removes the condition where
the fee token falls back to the validator fee token and sets the
`DEFAULT_FEE_TOKEN` (ie. first TIP20 deployed) as a fallback when no fee
token is found.

---------

Co-authored-by: Federico Gimenez <federico.gimenez@gmail.com>
Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>
Co-authored-by: Federico Gimenez <fgimenez@users.noreply.github.com>